requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
distlib==0.3.7
Django==3.0.3
django-bootstrap4==2.2.0
djangorestframework==3.14.0
factory-boy==3.0.1
Faker==21.0.0
filelock==3.13.1
numpy==1.26.2
pandas==2.1.4
pbr==6.0.0
platformdirs==4.1.0
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
